manajemen kinerja karyawan dan manajer

+++ app/Http/Controllers/TaskController.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use Carbon\Carbon;
 use App\Models\Task;
 use App\Models\User;
+use App\Models\SubTask;
+use App\Models\TipeTask;
 use App\Models\UsersTask;
 use Illuminate\Http\Request;
 use App\Models\ProjectPerusahaan;
-use App\Models\SubTask;
-use App\Models\TipeTask;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 
@@ -29,14 +30,26 @@ class TaskController extends Controller
                 'task.users_task.user.dataDiri.kepegawaian.subJabatan',
             ])
             ->get();
-        return view('task.daftar_task', compact('title', 'tasks', 'tipeTask', 'users', 'project','userTasks'));
+        return view('task.daftar_task', compact(
+            'title', 
+            'tasks', 
+            'tipeTask', 
+            'users', 
+            'project','userTasks'
+        ));
     }
     
     public function detail($id)
     {
         $title = 'Detail Task';
         $project = ProjectPerusahaan::all();
-        $task = Task::find($id);
+        $task = Task::with([
+            'project_perusahaan',
+            'tipe_task',
+            'users_task.user.dataDiri.kepegawaian.subJabatan',
+            'sub_task.lampiran'
+        ])
+        ->findOrFail($id);
         $tipeTask = TipeTask::all();
         $user = UsersTask::where('task_id', $id)
             ->with([
@@ -44,12 +57,51 @@ class TaskController extends Controller
                 'user.dataDiri.kepegawaian.subJabatan',
             ])
             ->get();
-        $existingUserIds = UsersTask::where('task_id', $id)->pluck('user_id')->toArray();
-        $users = User::whereNotIn('id', $existingUserIds)->get();
-        $subTask = SubTask::where('task_id', $id)->with(['lampiran'])->get();
+        $existingUserIds = UsersTask::where('task_id', $id)
+            ->pluck('user_id')
+            ->toArray();
+        $users = User::whereNotIn('id', $existingUserIds)
+            ->get();
+        $subTask = SubTask::where('task_id', $id)
+            ->with(['lampiran', 'user'])
+            ->get();
+        $subTaskUser = SubTask::where('task_id', $id)
+            ->where('user_id', Auth::user()->id)
+            ->with(['lampiran', 'user'])
+            ->get();
+        $totalSubtask = SubTask::where('task_id', $id)
+            ->count();
+        $approvedSubtask = SubTask::where('task_id', $id)
+            ->where('status', 'approve')
+            ->count();
+        $progressPercentage = $totalSubtask > 0 
+            ? round(($approvedSubtask / $totalSubtask) * 100) 
+            : 0;
+        if ($progressPercentage == 100 && $task->status != 'selesai') {
+            $task->update([
+                'status' => 'selesai',
+                'tgl_selesai' => now()
+            ]);
+            
+            $task = $task->fresh();
+        } elseif ($progressPercentage < 100 && $task->status == 'selesai') {
+            $task->update([
+                'status' => 'proses',
+                'tgl_selesai' => null
+            ]);
+        }
 
-
-        return view('task.detail_task', compact('title', 'task', 'user', 'users', 'tipeTask', 'project', 'subTask'));
+        return view('task.detail_task', compact(
+            'title', 
+            'task', 
+            'user', 
+            'users', 
+            'tipeTask', 
+            'project', 
+            'subTask', 
+            'subTaskUser',
+            'progressPercentage'
+        ));
     }
     public function store(Request $request)
     {
@@ -57,12 +109,16 @@ class TaskController extends Controller
             'tipe_task' => 'nullable',
             'project_perusahaan_id' => 'nullable',
             'nama_task' => 'required',
+            'tgl_task' => 'required',
+            'deadline_task' => 'nullable',
+            'status' => 'nullable',
             'keterangan' => 'nullable',
             'upload' => 'nullable|file|mimes:pdf,xls,xlsx,doc,docx,jpg,jpeg,png|max:5120',
             'user.*' => 'nullable',
         ]);
 
-        $tipeTask = TipeTask::where('slug', $request->tipe_task)->first();
+        $tipeTask = TipeTask::where('slug', $request->tipe_task)
+            ->first();
         if (!$tipeTask) {
             return redirect()->back()->with('error', 'Tipe Task tidak ditemukan!.');
         }
@@ -75,17 +131,30 @@ class TaskController extends Controller
             $file->move(public_path('uploads'), $fileName);
             $uploadPath = $fileName;
         }
+        
+        if ($request->tgl_task !== null) {
+            $waktuMulai = Carbon::parse($request->tgl_task);
+            $hariIni = Carbon::today();
+            if ($waktuMulai->lessThanOrEqualTo($hariIni)) {
+                $status = "proses";
+            } else {
+                $status = "belum";
+            }
+        } else {
+            $status = "belum";
+        }
 
         $data = [
             'tipe_tasks_id' => $tipeTask->id,
             'project_perusahaan_id' => $request->project_perusahaan_id,
-            'tgl_task' => now()->format('Y-m-d'),
+            'tgl_task' => $request->tgl_task,
+            'deadline' => $request->deadline_task,
+            'status' => $status,
             'user_id' => Auth::user()->id,
             'nama_task' => $request->nama_task,
             'keterangan' => $request->keterangan,
             'upload' => $uploadPath,
         ];
-
         $task = Task::create($data);
 
         if (Auth::check() && Auth::user()->role->slug == 'manager') {
@@ -111,6 +180,9 @@ class TaskController extends Controller
         $request->validate([
             'tipe_task' => 'nullable',
             'project_perusahaan_id' => 'nullable',
+            'tgl_task' => 'required',
+            'deadline_task' => 'required',
+            'status' => 'nullable',
             'nama_task' => 'required',
             'keterangan' => 'required',
             'upload' => 'nullable|file|mimes:pdf,xls,xlsx,doc,docx,jpg,jpeg,png,gif|max:5120',
@@ -137,6 +209,19 @@ class TaskController extends Controller
             $file->move(public_path('uploads'), $filename);
             $task->upload = $filename;
         }
+
+        
+        if ($request->tgl_task !== null) {
+            $waktuMulai = Carbon::parse($request->tgl_task);
+            $hariIni = Carbon::today();
+            if ($waktuMulai->lessThanOrEqualTo($hariIni)) {
+                $status = "proses";
+            } else {
+                $status = "belum";
+            }
+        } else {
+            $status = "belum";
+        }
         
         $data = [
             'tipe_tasks_id' => $tipeTask->id,
@@ -145,9 +230,61 @@ class TaskController extends Controller
             'nama_task' => $request->nama_task,
             'keterangan' => $request->keterangan,
             'upload' => $task->upload,
-            'tgl_task' => now()->format('Y-m-d'),
+            'tgl_task' => $request->tgl_task,
+            'deadline' => $request->deadline_task,
+            'status' => $status,
         ];
+        
+        $task->update($data);
+        return redirect()->back()->with('success', 'Task berhasil di update');
+    }
+    public function updateDetailTask(Request $request, $id)
+    {
+        $task = Task::find($id);
 
+        $request->validate([
+            'tipe_task' => 'nullable',
+            'nama_project' => 'nullable',
+            'tgl_task' => 'required',
+            'tgl_selesai' => 'nullable',
+            'deadline' => 'required',
+            'status' => 'nullable',
+            'nama_task' => 'required',
+        ]);
+        $tipeTask = TipeTask::where('slug', $request->tipe_task)->first();
+        if (!$tipeTask) {
+            return redirect()->back()->with('error', 'Tipe Task tidak ditemukan!.');
+        }
+        
+        if ($request->tgl_task !== null) {
+            $waktuMulai = Carbon::parse($request->tgl_task);
+            $hariIni = Carbon::today();
+            if ($waktuMulai->lessThanOrEqualTo($hariIni)) {
+                $status = "proses";
+            } else {
+                $status = "belum";
+            }
+        } else {
+            $status = "belum";
+        }
+
+        if ($request->filled('tgl_selesai')) {
+            $status = "selesai";
+        } else {
+            $status = ($request->status <= now()) ? "proses" : "belum";
+        }
+        
+        $data = [
+            'tipe_tasks_id' => $tipeTask->id,
+            'project_perusahaan_id' => $request->nama_project,
+            'user_id' => $request->user_id,
+            'nama_task' => $request->nama_task,
+            'tgl_task' => $request->tgl_task,
+            'tgl_selesai' => $request->tgl_selesai,
+            'deadline' => $request->deadline,
+            'status' => $status,
+        ];
+        
         $task->update($data);
         return redirect()->back()->with('success', 'Task berhasil di update');
     }
@@ -186,6 +323,7 @@ class TaskController extends Controller
 
         $request->validate([
             'upload' => 'nullable|file|mimes:pdf,xls,xlsx,doc,docx,jpg,jpeg,png,gif|max:5120',
+            'keterangan' => 'required',
         ]);
 
         if ($request->hasFile('upload')) {
@@ -202,18 +340,23 @@ class TaskController extends Controller
         }
 
         $task->save();
+        $task->update([
+            'keterangan' => $request->keterangan,
+        ]);
 
         return redirect()->back()->with('success', 'Lampiran berhasil di update');
     }
-
     public function destroy($id)
     {
         $task = Task::find($id);
+        if (!$task) {
+            return redirect()->back()->with('error', 'Task tidak ditemukan.');
+        }
+        $task->sub_task()->delete();
         $task->delete();
 
-        return redirect()->back()->with('success', 'Task berhasil di Hapus');
+        return redirect()->back()->with('success', 'Task berhasil dihapus.');
     }
-    
     public function destroyUserTask($id)
     {
         $deleted = UsersTask::where('id', $id)
